// schematics/src/container-component-pair/schema.d.ts
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Creates a new container and component.
 */
export interface PwaContainerComponentPairOptionsSchema {
  path?: string;
  tsext?: string;
  project?: string;
  prefix?: string;
  styleext?: string;
  selector?: string;
  module?: string;
  /**
   * The name of the container and component pair.
   */
  name?: string;
  /**
   * When true, generates and includes a style file for the container.
   */
  styleFile?: boolean;
}